Create TASK-2(Word count)
This Python code counts how many words are in a given text. It removes extra spaces at the beginning and end of the text, splits the text into words based on spaces, and then counts how many words there are. You can use it to find out how many words are in any piece of text you want to analyze.